Don't remove PR previews after closing the PR
This causes github queue issues when multiple PRs are closed one after
another. The later jobs cancel the jobs that are already in the queue
which in turn results in us missing some previews that can already be
removed.

This is also no longer necessary since we have a separate cleanup job
that runs nightly and can be triggered manually from the actions menu.

Signed-off-by: Tomasz Gorochowik <tgorochowik@antmicro.com>